Move Download to shared components.
