+++ src/shared/components/download/styles/download.scss
@@ -1,0 +1,11 @@
+.codeblock {
+  margin: 1rem;
+
+  &__container {
+    overflow: auto;
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    border: 1px solid black;
+  }
+}
