--- src/uniprotkb/components/download/styles/download.scss
@@ -1,18 +1,0 @@
-.preview {
-  margin: 1rem;
-
-  &__container {
-    overflow: hidden;
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    border: 1px solid black;
-    overflow-x: auto;
-    overflow-y: auto;
-  }
-
-  &__inner {
-    padding: 0.5rem;
-    height: 17.5rem;
-  }
-}
